+++ src/main/java/core/mvc/DispatcherServlet.java
@@ -1,9 +1,7 @@
-package core.mvc.asis;
+package core.mvc;
 
-import core.mvc.DefaultView;
-import core.mvc.ModelAndView;
-import core.mvc.tobe.AnnotationHandlerMapping;
-import core.mvc.tobe.HandlerExecution;
+import core.mvc.handlermapping.AnnotationHandlerMapping;
+import core.mvc.handlermapping.HandlerExecution;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -12,20 +10,16 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.util.Collections;
 
 @WebServlet(name = "dispatcher", urlPatterns = "/", loadOnStartup = 1)
 public class DispatcherServlet extends HttpServlet {
     private static final long serialVersionUID = 1L;
     private static final Logger logger = LoggerFactory.getLogger(DispatcherServlet.class);
 
-    private RequestMapping rm;
     private AnnotationHandlerMapping annotationMapping;
 
     @Override
     public void init() throws ServletException {
-        rm = new RequestMapping();
-        rm.initMapping();
         annotationMapping = new AnnotationHandlerMapping("next.controller");
         annotationMapping.initialize();
     }
@@ -36,10 +30,6 @@ public class DispatcherServlet extends HttpServlet {
         logger.debug("Method : {}, Request URI : {}", req.getMethod(), requestUri);
 
         HandlerExecution execution = annotationMapping.getHandler(req);
-        if (execution == null) {
-            serviceWithLegacy(req, resp, requestUri);
-            return;
-        }
         service(req, resp, execution);
     }
 
@@ -55,21 +45,5 @@ public class DispatcherServlet extends HttpServlet {
 
     private void logError(Throwable exception) {
         logger.error("Exception : {}", exception);
-    }
-
-    private void serviceWithLegacy(HttpServletRequest req, HttpServletResponse resp, String requestUri) throws ServletException {
-        Controller controller = rm.findController(requestUri);
-        try {
-            String viewName = controller.execute(req, resp);
-            move(viewName, req, resp);
-        } catch (Throwable e) {
-            logError(e);
-            throw new ServletException(e.getMessage());
-        }
-    }
-
-    private void move(String viewName, HttpServletRequest req, HttpServletResponse resp)
-            throws Exception {
-        new DefaultView(viewName).render(Collections.emptyMap(), req, resp);
     }
 }
