+++ src/main/java/next/controller/CreateUserController.java
@@ -4,8 +4,8 @@ import core.annotation.web.Controller;
 import core.annotation.web.RequestMapping;
 import core.annotation.web.RequestMethod;
 import core.db.DataBase;
-import core.mvc.DefaultView;
 import core.mvc.ModelAndView;
+import core.mvc.view.DefaultView;
 import next.model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
