+++ src/main/java/next/controller/ListUserController.java
@@ -4,8 +4,8 @@ import core.annotation.web.Controller;
 import core.annotation.web.RequestMapping;
 import core.annotation.web.RequestMethod;
 import core.db.DataBase;
-import core.mvc.DefaultView;
 import core.mvc.ModelAndView;
+import core.mvc.view.DefaultView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
