+++ src/main/java/next/controller/UpdateFormUserController.java
@@ -4,8 +4,8 @@ import core.annotation.web.Controller;
 import core.annotation.web.RequestMapping;
 import core.annotation.web.RequestMethod;
 import core.db.DataBase;
-import core.mvc.DefaultView;
 import core.mvc.ModelAndView;
+import core.mvc.view.DefaultView;
 import next.model.User;
 
 import javax.servlet.http.HttpServletRequest;
